+++ PowerShellForFOD/Public/Import-FODStaticScan.ps1
@@ -32,6 +32,8 @@ function Import-FODStaticScan {
         [Parameter(Mandatory)]
         [system.io.fileinfo]$ScanFile,
 
+        [switch]$Raw,
+
         [Parameter()]
         [ValidateNotNullOrEmpty()]
         [string]$Token = $Script:PS4FOD.Token,